+++ jarboot-common/src/main/java/com/mz/jarboot/common/CommandResponse.java
@@ -5,7 +5,6 @@ package com.mz.jarboot.common;
  * @author majianzheng
  */
 public class CommandResponse implements CmdProtocol {
-    private static final char SUCCESS_FLAG = '1';
     private Boolean success;
     private ResponseType responseType = ResponseType.UNKNOWN;
     private String body;
@@ -20,30 +19,35 @@ public class CommandResponse implements CmdProtocol {
         //格式: 前3个字符是控制位 + 数据体
         //控制位 0 响应类型、1 是否成功、2保留填-
         char rt = this.getResponseTypeChar();
+        if (Boolean.TRUE.equals(success)) {
+            rt = (char)(rt | CommandConst.SUCCESS_FLAG);
+        }
         return new StringBuilder()
-                //响应类型
+                //响应类型及是否成功的头
                 .append(rt)
-                //是否成功标志
-                .append(Boolean.TRUE.equals(success) ? SUCCESS_FLAG : '0')
-                //保留位
-                .append('-')
                 .append(body)
-                .append(' ')
+                .append(CommandConst.PROTOCOL_SPLIT)
                 //最后填充sessionId
                 .append(sessionId)
                 .toString();
     }
     @Override
     public void fromRaw(String raw) {
-        this.success = false;
+        char h = raw.charAt(0);
+        this.success = CommandConst.SUCCESS_FLAG == (CommandConst.SUCCESS_FLAG & h);
+        //取反再与得到真实响应类型
+        h = (char)(h & ~CommandConst.SUCCESS_FLAG);
         //反向解析出类实例
         //获取响应类型
-        switch (raw.charAt(0)) {
+        switch (h) {
             case CommandConst.ACK_TYPE:
                 this.setResponseType(ResponseType.ACK);
                 break;
             case CommandConst.ONLINE_TYPE:
                 this.setResponseType(ResponseType.ONLINE);
+                break;
+            case CommandConst.STD_OUT_TYPE:
+                this.setResponseType(ResponseType.STD_OUT);
                 break;
             case CommandConst.CONSOLE_TYPE:
                 this.setResponseType(ResponseType.CONSOLE);
@@ -59,15 +63,13 @@ public class CommandResponse implements CmdProtocol {
                 break;
         }
 
-        this.success = SUCCESS_FLAG == raw.charAt(1);
-
-        int l = raw.lastIndexOf(' ');
+        int l = raw.lastIndexOf(CommandConst.PROTOCOL_SPLIT);
         if (-1 == l) {
             this.success = false;
             this.body = "协议错误，未发现sessionId";
             return;
         }
-        this.body = raw.substring(3, l);
+        this.body = raw.substring(1, l);
         this.sessionId = raw.substring(l + 1);
     }
 
@@ -82,6 +84,8 @@ public class CommandResponse implements CmdProtocol {
                 return CommandConst.ACK_TYPE;
             case ONLINE:
                 return CommandConst.ONLINE_TYPE;
+            case STD_OUT:
+                return CommandConst.STD_OUT_TYPE;
             case CONSOLE:
                 return CommandConst.CONSOLE_TYPE;
             case JSON_RESULT: