+++ jarboot-common/src/main/java/com/mz/jarboot/common/ResponseType.java
@@ -9,22 +9,32 @@ public enum ResponseType {
      * 服务上线
      */
     ONLINE,
+
     /**
      * 命令应答
      */
     ACK,
+
     /**
-     * 控制太消息打印
+     * 程序中的标准输出流
+     */
+    STD_OUT,
+
+    /**
+     * 控制台消息打印
      */
     CONSOLE,
+
     /**
      * Json类型的执行结果
      */
     JSON_RESULT,
+
     /**
      * 执行执行完成
      */
     COMMAND_END,
+
     /**
      * 非法的未知类型
      */