+++ jarboot-core/src/main/java/com/mz/jarboot/core/cmd/impl/BytesCommand.java
@@ -1,5 +1,6 @@
 package com.mz.jarboot.core.cmd.impl;
 
+import com.alibaba.bytekit.utils.IOUtils;
 import com.alibaba.deps.org.objectweb.asm.ClassReader;
 import com.alibaba.deps.org.objectweb.asm.tree.AbstractInsnNode;
 import com.alibaba.deps.org.objectweb.asm.tree.ClassNode;
@@ -19,6 +20,7 @@ import org.slf4j.LoggerFactory;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * show the jvm detail
@@ -69,34 +71,30 @@ public class BytesCommand extends AbstractCommand {
             }
         }
         if (null == cls) {
-            session.end(true, "没有找到类," + this.classPattern);
+            session.end(true, "Not find," + this.classPattern);
             return;
         }
         //打印classloader
         session.console("ClassLoader: " + cls.getClassLoader().toString());
         session.console("------");
-        EnvironmentContext.getTransformerManager()
-                .addOnceTransformer(cls, (className, classfileBuffer) -> {
-                    try {
-                        ClassReader reader = new ClassReader(classfileBuffer);
-                        ClassNode classNode = new ClassNode();
-                        reader.accept(classNode, 0);
-                        final List<MethodNode> methods = classNode.methods;
-                        for (MethodNode m : methods) {
-                            InsnList inList = m.instructions;
-                            session.console(m.name);
-                            for (int i = 0; i < inList.size(); i++) {
-                                session.console(nodeToString(inList.get(i)));
-                            }
-                        }
-                    } catch (Exception e) {
-                        logger.warn(e.getMessage(), e);
-                        session.console("解析类失败，" + e.getMessage());
-                    }
-                    session.end();
-                });
-
-        EnvironmentContext.getTransformerManager().retransformClasses(cls);
+        try {
+            byte[] classfileBuffer = IOUtils.getBytes(Objects.requireNonNull(cls.getClassLoader()
+                    .getResourceAsStream(cls.getName().replace('.', '/') + ".class")));
+            ClassReader reader = new ClassReader(classfileBuffer);
+            ClassNode classNode = new ClassNode();
+            reader.accept(classNode, 0);
+            final List<MethodNode> methods = classNode.methods;
+            for (MethodNode m : methods) {
+                InsnList inList = m.instructions;
+                session.console(m.name);
+                for (int i = 0; i < inList.size(); i++) {
+                    session.console(nodeToString(inList.get(i)));
+                }
+            }
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+        }
+        session.end();
     }
 
     public static String nodeToString(AbstractInsnNode node){