+++ jarboot-server/src/main/java/com/mz/jarboot/utils/TaskUtils.java
@@ -4,12 +4,10 @@ import com.mz.jarboot.base.AgentManager;
 import com.mz.jarboot.common.OSUtils;
 import com.mz.jarboot.constant.CommonConst;
 import com.mz.jarboot.dto.ServerSettingDTO;
-import com.mz.jarboot.event.ApplicationContextUtils;
 import com.mz.jarboot.event.NoticeEnum;
 import com.mz.jarboot.ws.WebSocketManager;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.math.NumberUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.CollectionUtils;
@@ -25,25 +23,10 @@ import java.util.*;
  * @author majianzheng
  */
 public class TaskUtils {
-    private static volatile long startWaitTime = -1;
     private static final Logger logger = LoggerFactory.getLogger(TaskUtils.class);
     private static final String ADDER_ARGS_PREFIX = CommonConst.JARBOOT_NAME + CommonConst.DOT +
             CommonConst.SERVICES + CommonConst.DOT;
-    private static final int MIN_START_WAIT = 1500;
-    private static final int MAX_START_WAIT = 30000;
 
-    private static long getStartWaitTime() {
-        if (-1 == startWaitTime) {
-            synchronized (TaskUtils.class) {
-                String val = ApplicationContextUtils.getEnv("jarboot.start-wait-time", "5000");
-                startWaitTime = NumberUtils.toLong(val, 5000);
-                if (startWaitTime < MIN_START_WAIT || startWaitTime > MAX_START_WAIT) {
-                    startWaitTime = 15000;
-                }
-            }
-        }
-        return startWaitTime;
-    }
     /**
      * 检查服务进程是否存活
      * @param server 服务名
@@ -111,13 +94,15 @@ public class TaskUtils {
                 cmdBuilder.append(CommonConst.EXE_EXT);
             }
         }
-        cmdBuilder.append(StringUtils.SPACE);
-        // jvm 配置
-        if (StringUtils.isBlank(jvm)) {
-            cmdBuilder.append(CommonConst.ARG_JAR).append(jar);
-        } else {
-            cmdBuilder.append(jvm).append(CommonConst.ARG_JAR).append(jar);
-        }
+        cmdBuilder
+                .append(StringUtils.SPACE)
+                // jvm 配置
+                .append(jvm)
+                // Java agent
+                .append(SettingUtils.getAgentStartOption(server))
+                .append(StringUtils.SPACE)
+                // 待执行的jar
+                .append(CommonConst.ARG_JAR).append(jar);
 
         // 传入参数
         String startArg = setting.getArgs();
@@ -142,8 +127,9 @@ public class TaskUtils {
         //打印命令行
         WebSocketManager.getInstance().sendConsole(server, cmd);
         // 启动
-        startTask(cmd, setting.getEnv(), workHome,
-                text -> WebSocketManager.getInstance().sendConsole(server, text));
+        startTask(cmd, setting.getEnv(), workHome);
+        //等待启动完成，最长2分钟
+        AgentManager.getInstance().waitServerStarted(server, 120000);
     }
 
     /**
@@ -269,9 +255,8 @@ public class TaskUtils {
         pidList.add(Integer.parseInt(builder.toString()));
     }
 
-    public static void startTask(String command, String environment, String workHome, PushMsgCallback callback) {
+    public static void startTask(String command, String environment, String workHome) {
         String[] en;
-        final long waitTime = getStartWaitTime();
         if (StringUtils.isBlank(environment)) {
             en = null;
         } else {
@@ -282,22 +267,11 @@ public class TaskUtils {
             dir = new File(workHome);
         }
 
-        String msg = "Finished.";
-        try (InputStream inputStream = Runtime.getRuntime().exec(command, en, dir).getInputStream()) {
-            if (null == callback) {
-                return;
-            }
-            intervalReadStream(callback, waitTime, inputStream);
-        } catch (InterruptedException e) {
-            msg = e.getMessage();
-            logger.error(msg, e);
-            Thread.currentThread().interrupt();
+        try {
+            Runtime.getRuntime().exec(command, en, dir);
         } catch (Exception e) {
-            msg = e.getMessage();
-            logger.error(msg, e);
-        }
-        if (null != callback) {
-            callback.sendMessage(msg);
+            logger.error(e.getMessage(), e);
+            WebSocketManager.getInstance().notice("Start task error " + e.getMessage(), NoticeEnum.ERROR);
         }
     }